# src/main/resources/openapi/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
tags:
- name: Exchange Rates
  description: Operations related to exchange rate conversions
paths:
  /api/exchange/convert:
    post:
      tags:
      - Exchange Rates
      summary: Convert currency
      description: Converts an amount from one currency to another.
      operationId: convert
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversionRequest"
        required: true
      responses:
        "200":
          description: Conversion successful
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConversionResponse"
        "404":
          description: Currency code not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/exchange/rates:
    get:
      tags:
      - Exchange Rates
      summary: Get all exchange rates
      description: Returns a list of supported exchange rates.
      operationId: getExchangeRates
      responses:
        "200":
          description: List of exchange rates
          content:
            '*/*':
              schema:
                type: array
                $ref: "#/components/schemas/ExchangeRatesResponse"
components:
  schemas:
    ConversionResponse:
      type: object
      description: Response object containing the result of a currency conversion.
      properties:
        fromCurrency:
          type: string
          description: The original currency code provided by the client.
          example: USD
        toCurrency:
          type: string
          description: The target currency code to which the amount is converted.
          example: EUR
        originalAmount:
          type: number
          description: The original amount requested for conversion.
          example: 100.0
        convertedAmount:
          type: number
          description: The converted amount in the target currency.
          example: 85.67
    ErrorResponse:
      type: object
      description: Represents an error response returned by the API
      properties:
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the error occurred
          example: 2025-08-30T12:34:56.789
        status:
          type: integer
          format: int32
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Short HTTP status description
          example: Bad Request
        errorMessages:
          type: array
          description: "List of validation errors, if any"
          items:
            type: string
    ConversionRequest:
      type: object
      properties:
        fromCurrency:
          type: string
          description: Currency code you want to convert from
          example: USD
          pattern: "^[A-Z]{3}$"
        toCurrency:
          type: string
          description: Currency code you want to convert to
          example: EUR
          pattern: "^[A-Z]{3}$"
        amount:
          type: number
          description: Amount to convert
          example: 100.5
      required:
      - amount
      - fromCurrency
      - toCurrency
    ExchangeRatesResponse:
      type: object
      description: Response object representing an exchange rate for a given currency.
      properties:
        currency:
          type: string
          description: The 3-letter ISO currency code.
          example: USD
        rate:
          type: number
          description: The exchange rate of the currency against EUR.
          example: 1.1676